moved internals/spec to pkg/spec
