+++ pkg/spec/extension.go
