+++ pkg/spec/extension_test.go
